pyproject.toml
[tool.poetry]
name = "gpt-dev"
version = "0.1.0"
description = ""
authors = ["Shruti Patel <shruti222patel@gmail.com>"]
readme = "README.md"
packages = [{include = "gpt_dev"}]

[tool.poetry.dependencies]
python = "^3.10"
openai = "^0.27.8"
tenacity = "^8.2.2"
pygithub = "^1.59.0"
requests = "^2.31.0"
unidiff = "^0.7.5"

[tool.poetry.group.dev.dependencies]
pytest = "^7.3.2"
pre-commit = "^3.3.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
